requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-chroma 
langchain-community 
langchain-openai
chroma